model/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-debug-toolbar==4.2.0
django-import-export==3.3.1
django-simpleui==2023.8.28
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.6
xlrd==2.0.1
xlwt==1.3.0
